# System.Xml.NameTable optimizations 