@kronic 2021-03-22
|       Method |      Mean |     Error |    StdDev |
|------------- |----------:|----------:|----------:|
| After|  9.645 ns | 0.0086 ns | 0.0076 ns |
| Before | 13.176 ns | 0.0080 ns | 0.0067 ns |