@kronic 2021-03-22
I will fix it. Is it worth it to also inline the ComputeHash32 method?